Routes updated to post and get so put and delete are not blocked

+++ app/Repositories/ProfileRepository.php
@@ -13,17 +13,25 @@ class ProfileRepository
     {
 
         DB::transaction(function () use ($data) {
-            $image = $data['image'];
-            $imageHeader = $data['image_header'];
-            $this->uploads($image, auth()->user()->email . '/profile/');
-            $this->uploads($imageHeader, auth()->user()->email . '/profile/');
+
+            $image = null;
+            $imageHeader = null;
+
+            if (data_get($data, 'image')) {
+                $image = $data['image']->hashName();
+                $this->uploads($image, auth()->user()->email . '/profile/');
+            }
+            if (data_get($data, 'image_header')) {
+                $imageHeader = $data['image_header']->hashName();
+                $this->uploads($imageHeader, auth()->user()->email . '/profile/');
+            }
 
             $created = Profile::create([
                 'day_of_birth' => data_get($data, 'day_of_birth'),
                 'gender' => data_get($data, 'gender'),
                 'country_id' => data_get($data, 'country_id'),
-                'image' => $image->hashName(),
-                'image_header' => $imageHeader->hashName(),
+                'image' => $image,
+                'image_header' => $imageHeader,
                 'title' => data_get($data, 'title'),
                 'bio' => data_get($data, 'bio'),
                 'likes' => data_get($data, 'likes'),
@@ -44,19 +52,40 @@ class ProfileRepository
     public function update(array $data, Profile $profile)
     {
         DB::transaction(function () use ($data, $profile) {
-            $image = $data['image'];
-            $imageHeader = $data['image_header'];
 
-            $this->removeFile(auth()->user()->email . '/profile/' . $profile->image);
-            $this->removeFile(auth()->user()->email . '/profile/' . $profile->image_header);
+            $imageSource = $profile->image;
+            $imageHeaderSource = $profile->image_header;
+
+            if (data_get($data, 'image') && is_object($data['image'])) {
+                $image = data_get($data, 'image');
+                $imageSource = $data['image']->hashName();
+                $this->uploads($image, auth()->user()->email . '/profile/');
+                $this->removeFile(auth()->user()->email . '/profile/' . $profile->image);
+            }
+            if (data_get($data, 'image_header') && is_object($data['image_header'])) {
+                $imageHeader = data_get($data, 'image_header');
+                $imageHeaderSource = $data['image_header']->hashName();
+                $this->uploads($imageHeader, auth()->user()->email . '/profile/');
+                $this->removeFile(auth()->user()->email . '/profile/' . $profile->image_header);
+            }
+
             $updated = $profile->update([
-                'image' => $image->hashName(),
-                'image_header' => $imageHeader->hashName(),
-                ...$data
+                'image' => $imageSource,
+                'image_header' => $imageHeaderSource,
+                'day_of_birth' => data_get($data, 'day_of_birth'),
+                'gender' => data_get($data, 'gender'),
+                'country_id' => data_get($data, 'country_id'),
+                'title' => data_get($data, 'title'),
+                'bio' => data_get($data, 'bio'),
+                'likes' => data_get($data, 'likes'),
+                'dislikes' => data_get($data, 'dislikes'),
+                'address' => data_get($data, 'address'),
+                'phone' => data_get($data, 'phone'),
+                'public_email' => data_get($data, 'public_email'),
+                'level_id' => data_get($data, 'level_id'),
+                'sentimental_id' => data_get($data, 'sentimental_id'),
             ]);
 
-            $this->uploads($image, auth()->user()->email . '/profile/');
-            $this->uploads($imageHeader, auth()->user()->email . '/profile/');
             throw_if(!$updated, \Exception::class,  'Error updating profile');
         });
         return $profile->refresh();
